agro_site/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
Django==2.2.16
djangorestframework==3.12.4
idna==3.3
Pillow==9.0.1
python-http-client==3.3.6
pytz==2021.3
requests==2.26.0
sendgrid==6.9.6
sorl-thumbnail==12.7.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
urllib3==1.26.8
django-phonenumber-field==5.0.0
phonenumbers==8.12.43
django-ratings==0.3.7
django-bootstrap3==21.2